refactor(release-analysis): extract Monte Carlo utils to shared module
The gammaSample and boxMullerNormal functions were duplicated across
MonteCarloChart, ReleaseVersionGroup, and ProductReleaseCard. Extract
them to utils/monteCarlo.js so all three components import from one
source.
