+++ modules/release-analysis/client/components/MonteCarloChart.vue
@@ -60,6 +60,8 @@ import {
   Title
 } from 'chart.js'
 
+import { gammaSample } from '../utils/monteCarlo'
+
 ChartJS.register(CategoryScale, LinearScale, BarElement, Tooltip, Title)
 
 const ITERATIONS = 1000
@@ -73,38 +75,6 @@ const props = defineProps({
   codeFreezeDate: { type: String, default: null },
   releaseDate: { type: String, default: null }
 })
-
-// ── Random sampling ──
-
-function boxMullerNormal() {
-  const u1 = Math.random()
-  const u2 = Math.random()
-  return Math.sqrt(-2 * Math.log(u1)) * Math.cos(2 * Math.PI * u2)
-}
-
-/**
- * Gamma(shape, scale) via Marsaglia & Tsang.
- * Used to model the waiting time for `shape` Poisson events at rate 1/scale.
- */
-function gammaSample(shape, scale) {
-  if (shape < 1) {
-    return gammaSample(shape + 1, scale) * Math.pow(Math.random(), 1 / shape)
-  }
-  const d = shape - 1 / 3
-  const c = 1 / Math.sqrt(9 * d)
-  for (let iter = 0; iter < 1000; iter++) {
-    let x, v
-    do {
-      x = boxMullerNormal()
-      v = 1 + c * x
-    } while (v <= 0)
-    v = v * v * v
-    const u = Math.random()
-    if (u < 1 - 0.0331 * x * x * x * x) return d * v * scale
-    if (Math.log(u) < 0.5 * x * x + d * (1 - v + Math.log(v))) return d * v * scale
-  }
-  return shape * scale
-}
 
 // ── Date helpers ──
